+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 45
+
+- Replace date visibility for date format
+- Fix accessibility checkbox visibility for API 37
+- Self heal orphan groups
+- Private space visibility
+- Fix wallpaper overlay stacking and list ghosting
+- Made monochrome icons bigger
+- Fix back gesture on main screen
+- Fix render issues on icon pack selection
+
 ## 44
 
 - Fix color picker